02_productivity/shiny_app/requirements.txt
shiny>=0.9.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
matplotlib>=3.8.0

